% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreBackgroundFunctionalityMGMT2340.R
\name{CUDmean.KSP}
\alias{CUDmean.KSP}
\title{Cumulative Uniform Distribution expected value (mean / arithimatic mean)}
\usage{
CUDmean.KSP(Top, Bottom)
}
\arguments{
\item{Top}{the Greater number in the range}

\item{Bottom}{the lesser number in the range}
}
\value{
expected value (or mean) of the range
}
\description{
Cumulative Uniform Distribution expected value (mean / arithimatic mean)
}
\examples{
A train usually arives between 30 to 200 seconds late what is the expected value CUDmean.KBS(200, 30)
}
