% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreBackgroundFunctionalityMGMT2340.R
\name{CUDprob.KSP}
\alias{CUDprob.KSP}
\title{Cumulative Unifrom Distribution Probability P(X < x) calculator}
\usage{
CUDprob.KSP(X, Top, Bottom)
}
\arguments{
\item{X}{value for probability P(X < x)}

\item{Top}{the Greater number in the range}

\item{Bottom}{the lesser number in the range}
}
\value{
probabiliy that X < x
}
\description{
Cumulative Unifrom Distribution Probability P(X < x) calculator
}
\examples{
A train usually arives between 30 to 200 seconds late what is the probability the train arives within 160 seconds: CUDprob.KBS(160, 200, 30)
}
