% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsAddedForHW.R
\name{CUDsd.KSP}
\alias{CUDsd.KSP}
\title{Cumulative Uniform Distribution Standard Deviation}
\usage{
CUDsd.KSP(Top, Bottom)
}
\arguments{
\item{Top}{the Greater number in the range}

\item{Bottom}{the lesser number in the range}
}
\value{
standard deviation of the range
}
\description{
Cumulative Uniform Distribution Standard Deviation
}
\examples{
A train usually arives between 30 to 200 seconds late what is the standard deviation CUDsd.KBS(200, 30)
}
