% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationStatistics.R
\name{killpack.return.calc}
\alias{killpack.return.calc}
\title{The average (geomean) rate of anual growth}
\usage{
killpack.return.calc(x)
}
\arguments{
\item{x}{a vector of intigers representing percentage growth written in decimal form (eg 1 percent written 0.01)}
}
\value{
the geometric rate of return accross all observations
}
\description{
Finds the average rate of return for a set number of periods even with negative intigers
}
\examples{
 killpack.return.calc(ExcelImport$column), killpack.return.calc(c(0.1384,0.1125,-0.1234,0.5432))
}
