% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreBackgroundFunctionalityMGMT2340.R
\name{LETSTUDYc1}
\alias{LETSTUDYc1}
\title{Chapter 2 Easy Starter}
\usage{
LETSTUDYc1()
}
\value{
downloads all files necessary for chapter 2 as well as loads all packages required
}
\description{
Chapter 2 Easy Starter
}
\examples{
LETSTUDYc1()
}
