% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationStatistics.R
\name{pop.geoMean}
\alias{pop.geoMean}
\title{Geometric Mean of a Population}
\usage{
pop.geoMean(x)
}
\arguments{
\item{x}{is a vector of intigers to run through the geometric mean formula}
}
\value{
The mean of the vector
}
\description{
Finds the geometric mean of a population
}
