% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationStatistics.R
\name{pop.sd}
\alias{pop.sd}
\title{Standard Deviation of Entire Population (N)}
\usage{
pop.sd(x)
}
\arguments{
\item{x}{A vector of intigers from which to calculate the standard deviation}
}
\value{
The standard deviation of a population
}
\description{
Takes in an entire set of values representing the full population and finds the standard deviation without removing sample bias (n-1)
}
