% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationStatistics.R
\name{pop.var}
\alias{pop.var}
\title{Variance of a Population}
\usage{
pop.var(x)
}
\arguments{
\item{x}{a vector of intigers from which to find the variance}
}
\value{
the variance of the Population (x)
}
\description{
Finds the variance of a whole population (N) removing the bias of sample sets (n-1)
}
