% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsAddedForHW.R
\name{X2YmLND}
\alias{X2YmLND}
\title{µx to µy lognormal function}
\usage{
X2YmLND(m, sd)
}
\arguments{
\item{m}{enter the log mean µx}

\item{sd}{enter the log sd aka øx}
}
\value{
µy of the lognormal µx
}
\description{
µx to µy lognormal function
}
\examples{
X2YmLND(6, 3)
}
