% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreBackgroundFunctionalityMGMT2340.R
\name{X2YsdLND}
\alias{X2YsdLND}
\title{øx to øy lognormal function}
\usage{
X2YsdLND(m, sd)
}
\arguments{
\item{m}{enter the log mean µx}

\item{sd}{enter the log sd aka øx}
}
\value{
øy of the lognormal øx
}
\description{
øx to øy lognormal function
}
\examples{
X2YmLND(6, 3)
}
