% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsAddedForHW.R
\name{Y2XmLND}
\alias{Y2XmLND}
\title{µy to µx lognormal function}
\usage{
Y2XmLND(m, sd)
}
\arguments{
\item{m}{enter the log mean µy}

\item{sd}{enter the log sd aka øy}
}
\value{
µx
}
\description{
µy to µx lognormal function
}
\examples{
X2YmLND(6, 3)
}
