% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreBackgroundFunctionalityMGMT2340.R
\name{Y2XsdLND}
\alias{Y2XsdLND}
\title{øx to øy lognormal function}
\usage{
Y2XsdLND(my, sdy)
}
\arguments{
\item{my}{enter the log mean sks µy}

\item{sdy}{yenter the log sd aka øy}
}
\value{
øy
}
\description{
øx to øy lognormal function
}
\examples{
X2YmLND(6, 3)
}
