% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreBackgroundFunctionalityMGMT2340.R
\name{zscore.KSP}
\alias{zscore.KSP}
\title{find Find z for P(Z < z)}
\usage{
zscore.KSP(x, m, sd)
}
\arguments{
\item{x}{value in question}

\item{m}{mean of the normal distribution}

\item{sd}{sd of the normal distribution}
}
\value{
zscore to look up probability on ztable from
}
\description{
find Find z for P(Z < z)
}
\examples{
X2YmLND(6, 3)
}
